services/common/requirements.txt
boto3
fastapi
psycopg2-binary
pydantic
pyjwt
python-json-logger
sqlalchemy
